@@ -26,13 +26,8 @@ import android.graphics.Canvas;
 import android.graphics.Paint;
 import android.graphics.Point;
 import android.graphics.Rect;
+import android.graphics.RectF;
 import android.graphics.Typeface;
-import android.support.annotation.ColorInt;
-import android.support.annotation.IntDef;
-import android.support.annotation.Keep;
-import android.support.v4.view.animation.FastOutLinearInInterpolator;
-import android.support.v4.view.animation.LinearOutSlowInInterpolator;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.ViewConfiguration;
@@ -43,8 +38,17 @@ import com.simplecityapps.recyclerview_fastscroll.utils.Utils;
 
 import java.lang.annotation.Retention;
 
+import androidx.annotation.ColorInt;
+import androidx.annotation.IntDef;
+import androidx.annotation.Keep;
+import androidx.annotation.NonNull;
+import androidx.interpolator.view.animation.FastOutLinearInInterpolator;
+import androidx.interpolator.view.animation.LinearOutSlowInInterpolator;
+import androidx.recyclerview.widget.RecyclerView;
+
 import static java.lang.annotation.RetentionPolicy.SOURCE;
 
+@SuppressWarnings("WeakerAccess")
 public class FastScroller {
     private static final int DEFAULT_AUTO_HIDE_DELAY = 1500;
 
@@ -52,10 +56,11 @@ public class FastScroller {
     private FastScrollPopup mPopup;
 
     private int mThumbHeight;
-    private int mWidth;
-
+    private int mThumbWidth;
     private Paint mThumb;
+
     private Paint mTrack;
+    private int mTrackWidth;
 
     private Rect mTmpRect = new Rect();
     private Rect mInvalidateRect = new Rect();
@@ -83,9 +88,18 @@ public class FastScroller {
     private int mThumbInactiveColor = 0x79000000;
     private boolean mThumbInactiveState;
 
+    private int mTouchSlop;
+
+    private int mLastY;
+
     @Retention(SOURCE)
-    @IntDef({FastScrollerPopupPosition.ADJACENT, FastScrollerPopupPosition.CENTER})
-    public @interface FastScrollerPopupPosition {
+    @IntDef({PopupTextVerticalAlignmentMode.TEXT_BOUNDS, PopupTextVerticalAlignmentMode.FONT_METRICS})
+    public @interface PopupTextVerticalAlignmentMode {
+        int TEXT_BOUNDS = 0;
+        int FONT_METRICS = 1;
+    }
+    @IntDef({PopupPosition.ADJACENT, PopupPosition.CENTER})
+    public @interface PopupPosition {
         int ADJACENT = 0;
         int CENTER = 1;
     }
@@ -97,13 +111,16 @@ public class FastScroller {
         mRecyclerView = recyclerView;
         mPopup = new FastScrollPopup(resources, recyclerView);
 
-        mThumbHeight = Utils.toPixels(resources, 48);
-        mWidth = Utils.toPixels(resources, 8);
+        mThumbHeight = Utils.toPixels(resources, 52);
+        mThumbWidth = Utils.toPixels(resources, 8);
+        mTrackWidth = Utils.toPixels(resources, 6);
 
         mTouchInset = Utils.toPixels(resources, -24);
 
         mThumb = new Paint(Paint.ANTI_ALIAS_FLAG);
         mTrack = new Paint(Paint.ANTI_ALIAS_FLAG);
+
+        mTouchSlop = ViewConfiguration.get(context).getScaledTouchSlop();
 
         TypedArray typedArray = context.getTheme().obtainStyledAttributes(
                 attrs, R.styleable.FastScrollRecyclerView, 0, 0);
@@ -117,9 +134,15 @@ public class FastScroller {
             int trackColor = typedArray.getColor(R.styleable.FastScrollRecyclerView_fastScrollTrackColor, 0x28000000);
             int popupBgColor = typedArray.getColor(R.styleable.FastScrollRecyclerView_fastScrollPopupBgColor, 0xff000000);
             int popupTextColor = typedArray.getColor(R.styleable.FastScrollRecyclerView_fastScrollPopupTextColor, 0xffffffff);
-            int popupTextSize = typedArray.getDimensionPixelSize(R.styleable.FastScrollRecyclerView_fastScrollPopupTextSize, Utils.toScreenPixels(resources, 44));
-            int popupBackgroundSize = typedArray.getDimensionPixelSize(R.styleable.FastScrollRecyclerView_fastScrollPopupBackgroundSize, Utils.toPixels(resources, 88));
-            @FastScrollerPopupPosition int popupPosition = typedArray.getInteger(R.styleable.FastScrollRecyclerView_fastScrollPopupPosition, FastScrollerPopupPosition.ADJACENT);
+            int popupTextSize = typedArray.getDimensionPixelSize(R.styleable.FastScrollRecyclerView_fastScrollPopupTextSize, Utils.toScreenPixels(resources, 32));
+            int popupBackgroundSize = typedArray.getDimensionPixelSize(R.styleable.FastScrollRecyclerView_fastScrollPopupBackgroundSize, Utils.toPixels(resources, 62));
+            @PopupTextVerticalAlignmentMode int popupTextVerticalAlignmentMode = typedArray.getInteger(R.styleable.FastScrollRecyclerView_fastScrollPopupTextVerticalAlignmentMode, PopupTextVerticalAlignmentMode.TEXT_BOUNDS);
+            @PopupPosition int popupPosition = typedArray.getInteger(R.styleable.FastScrollRecyclerView_fastScrollPopupPosition, PopupPosition.ADJACENT);
+
+            mThumbWidth =
+                typedArray.getDimensionPixelSize(R.styleable.FastScrollRecyclerView_fastScrollThumbWidth, mThumbWidth);
+            mTrackWidth =
+                typedArray.getDimensionPixelSize(R.styleable.FastScrollRecyclerView_fastScrollTrackWidth, mTrackWidth);
 
             mTrack.setColor(trackColor);
             mThumb.setColor(mThumbInactiveState ? mThumbInactiveColor : mThumbActiveColor);
@@ -127,6 +150,7 @@ public class FastScroller {
             mPopup.setTextColor(popupTextColor);
             mPopup.setTextSize(popupTextSize);
             mPopup.setBackgroundSize(popupBackgroundSize);
+            mPopup.setPopupTextVerticalAlignmentMode(popupTextVerticalAlignmentMode);
             mPopup.setPopupPosition(popupPosition);
         } finally {
             typedArray.recycle();
@@ -139,7 +163,7 @@ public class FastScroller {
                     if (mAutoHideAnimator != null) {
                         mAutoHideAnimator.cancel();
                     }
-                    mAutoHideAnimator = ObjectAnimator.ofInt(FastScroller.this, "offsetX", (Utils.isRtl(mRecyclerView.getResources()) ? -1 : 1) * mWidth);
+                    mAutoHideAnimator = ObjectAnimator.ofInt(FastScroller.this, "offsetX", (Utils.isRtl(mRecyclerView.getResources()) ? -1 : 1) * getWidth());
                     mAutoHideAnimator.setInterpolator(new FastOutLinearInInterpolator());
                     mAutoHideAnimator.setDuration(200);
                     mAutoHideAnimator.start();
@@ -149,7 +173,7 @@ public class FastScroller {
 
         mRecyclerView.addOnScrollListener(new RecyclerView.OnScrollListener() {
             @Override
-            public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+            public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
                 super.onScrolled(recyclerView, dx, dy);
 
                 if (!mRecyclerView.isInEditMode()) {
@@ -168,7 +192,7 @@ public class FastScroller {
     }
 
     public int getWidth() {
-        return mWidth;
+        return Math.max(mTrackWidth, mThumbWidth);
     }
 
     public boolean isDragging() {
@@ -181,8 +205,6 @@ public class FastScroller {
      */
     public void handleTouchEvent(MotionEvent ev, int downX, int downY, int lastY,
                                  OnFastScrollStateChangeListener stateChangeListener) {
-        ViewConfiguration config = ViewConfiguration.get(mRecyclerView.getContext());
-
         int action = ev.getAction();
         int y = (int) ev.getY();
         switch (action) {
@@ -194,7 +216,7 @@ public class FastScroller {
             case MotionEvent.ACTION_MOVE:
                 // Check if we should start scrolling
                 if (!mIsDragging && isNearPoint(downX, downY) &&
-                        Math.abs(y - downY) > config.getScaledTouchSlop()) {
+                        Math.abs(y - downY) > mTouchSlop) {
                     mRecyclerView.getParent().requestDisallowInterceptTouchEvent(true);
                     mIsDragging = true;
                     mTouchOffset += (lastY - downY);
@@ -207,19 +229,30 @@ public class FastScroller {
                     }
                 }
                 if (mIsDragging) {
-                    // Update the fastscroller section name at this touch position
-                    int top = 0;
-                    int bottom = mRecyclerView.getHeight() - mThumbHeight;
-                    float boundedY = (float) Math.max(top, Math.min(bottom, y - mTouchOffset));
-                    String sectionName = mRecyclerView.scrollToPositionAtProgress((boundedY - top) / (bottom - top));
-                    mPopup.setSectionName(sectionName);
-                    mPopup.animateVisibility(!sectionName.isEmpty());
-                    mRecyclerView.invalidate(mPopup.updateFastScrollerBounds(mRecyclerView, mThumbPosition.y));
+                    if (mLastY == 0 || Math.abs(mLastY - y) >= mTouchSlop) {
+                        mLastY = y;
+                        // Update the fastscroller section name at this touch position
+                        boolean layoutManagerReversed = mRecyclerView.isLayoutManagerReversed();
+                        int bottom = mRecyclerView.getHeight() - mThumbHeight;
+                        float boundedY = (float) Math.max(0, Math.min(bottom, y - mTouchOffset));
+
+                        // Represents the amount the thumb has scrolled divided by its total scroll range
+                        float touchFraction = boundedY / bottom;
+                        if (layoutManagerReversed) {
+                            touchFraction = 1 - touchFraction;
+                        }
+
+                        String sectionName = mRecyclerView.scrollToPositionAtProgress(touchFraction);
+                        mPopup.setSectionName(sectionName);
+                        mPopup.animateVisibility(!sectionName.isEmpty());
+                        mRecyclerView.invalidate(mPopup.updateFastScrollerBounds(mRecyclerView, mThumbPosition.y));
+                    }
                 }
                 break;
             case MotionEvent.ACTION_UP:
             case MotionEvent.ACTION_CANCEL:
                 mTouchOffset = 0;
+                mLastY = 0;
                 if (mIsDragging) {
                     mIsDragging = false;
                     mPopup.animateVisibility(false);
@@ -234,6 +267,8 @@ public class FastScroller {
         }
     }
 
+    RectF rect = new RectF();
+
     public void draw(Canvas canvas) {
 
         if (mThumbPosition.x < 0 || mThumbPosition.y < 0) {
@@ -241,20 +276,35 @@ public class FastScroller {
         }
 
         //Background
-        canvas.drawRect(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mWidth, mRecyclerView.getHeight() + mOffset.y, mTrack);
+        rect.set(mThumbPosition.x + mOffset.x + (mThumbWidth - mTrackWidth),
+                mOffset.y + mRecyclerView.getPaddingTop(),
+                mThumbPosition.x + mOffset.x + mTrackWidth + (mThumbWidth - mTrackWidth),
+                mRecyclerView.getHeight() + mOffset.y - mRecyclerView.getPaddingBottom());
+        canvas.drawRoundRect(rect,
+                mTrackWidth,
+                mTrackWidth,
+                mTrack);
 
         //Handle
-        canvas.drawRect(mThumbPosition.x + mOffset.x, mThumbPosition.y + mOffset.y, mThumbPosition.x + mOffset.x + mWidth, mThumbPosition.y + mOffset.y + mThumbHeight, mThumb);
+        rect.set(mThumbPosition.x + mOffset.x + (mThumbWidth - mTrackWidth) / 2,
+                mThumbPosition.y + mOffset.y,
+                mThumbPosition.x + mOffset.x + mThumbWidth + (mThumbWidth - mTrackWidth) / 2,
+                mThumbPosition.y + mOffset.y + mThumbHeight);
+        canvas.drawRoundRect(rect,
+                mThumbWidth,
+                mThumbWidth,
+                mThumb);
 
         //Popup
         mPopup.draw(canvas);
+
     }
 
     /**
      * Returns whether the specified points are near the scroll bar bounds.
      */
     private boolean isNearPoint(int x, int y) {
-        mTmpRect.set(mThumbPosition.x, mThumbPosition.y, mThumbPosition.x + mWidth,
+        mTmpRect.set(mThumbPosition.x, mThumbPosition.y, mThumbPosition.x + mTrackWidth,
                 mThumbPosition.y + mThumbHeight);
         mTmpRect.inset(mTouchInset, mTouchInset);
         return mTmpRect.contains(x, y);
@@ -265,9 +315,9 @@ public class FastScroller {
             return;
         }
         // do not create new objects here, this is called quite often
-        mInvalidateRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mWidth, mRecyclerView.getHeight() + mOffset.y);
+        mInvalidateRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mTrackWidth, mRecyclerView.getHeight() + mOffset.y);
         mThumbPosition.set(x, y);
-        mInvalidateTmpRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mWidth, mRecyclerView.getHeight() + mOffset.y);
+        mInvalidateTmpRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mTrackWidth, mRecyclerView.getHeight() + mOffset.y);
         mInvalidateRect.union(mInvalidateTmpRect);
         mRecyclerView.invalidate(mInvalidateRect);
     }
@@ -278,9 +328,9 @@ public class FastScroller {
             return;
         }
         // do not create new objects here, this is called quite often
-        mInvalidateRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mWidth, mRecyclerView.getHeight() + mOffset.y);
+        mInvalidateRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mTrackWidth, mRecyclerView.getHeight() + mOffset.y);
         mOffset.set(x, y);
-        mInvalidateTmpRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mWidth, mRecyclerView.getHeight() + mOffset.y);
+        mInvalidateTmpRect.set(mThumbPosition.x + mOffset.x, mOffset.y, mThumbPosition.x + mOffset.x + mTrackWidth, mRecyclerView.getHeight() + mOffset.y);
         mInvalidateRect.union(mInvalidateTmpRect);
         mRecyclerView.invalidate(mInvalidateRect);
     }
@@ -341,6 +391,7 @@ public class FastScroller {
     }
 
     public void setThumbColor(@ColorInt int color) {
+        mThumbActiveColor = color;
         mThumb.setColor(color);
         mRecyclerView.invalidate(mInvalidateRect);
     }
@@ -382,7 +433,7 @@ public class FastScroller {
         }
     }
 
-    public void setPopupPosition(@FastScrollerPopupPosition int popupPosition) {
+    public void setPopupPosition(@PopupPosition int popupPosition) {
         mPopup.setPopupPosition(popupPosition);
     }
 
